@@ -1,13 +1,11 @@
 package com.app.chao.chaoapp.contract.impl;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.app.chao.chaoapp.base.RxPresenter;
 import com.app.chao.chaoapp.bean.VideoInfo;
 import com.app.chao.chaoapp.bean.VideoRes;
 import com.app.chao.chaoapp.contract.VideoInfoContract;
-import com.app.chao.chaoapp.net.RetrofitHelper;
-import com.app.chao.chaoapp.net.VideoHttpResponse;
 import com.app.chao.chaoapp.utils.RxUtil;
 
 import org.simple.eventbus.EventBus;
@@ -16,7 +14,6 @@ import java.util.concurrent.TimeUnit;
 
 import rx.Observable;
 import rx.Subscription;
-import rx.functions.Action0;
 import rx.functions.Action1;
 
 public class VideoInfoPresenter extends RxPresenter implements VideoInfoContract.Presenter {
@@ -41,31 +38,31 @@ public class VideoInfoPresenter extends RxPresenter implements VideoInfoContract
 
     @Override
     public void getDetailData(String dataId) {
-        Subscription rxSubscription = RetrofitHelper.getVideoApi().getVideoInfo(dataId)
-                .compose(RxUtil.<VideoHttpResponse<VideoRes>>rxSchedulerHelper())
-                .compose(RxUtil.<VideoRes>handleResult())
-                .subscribe(new Action1<VideoRes>() {
-                    @Override
-                    public void call(final VideoRes res) {
-                        if (res != null) {
-                            mView.showContent(res);
-                            result = res;
-                            postData();
-//                            insertRecord();
-                        }
-                    }
-                }, new Action1<Throwable>() {
-                    @Override
-                    public void call(Throwable throwable) {
-                        mView.showError("数据加载失败");
-                    }
-                }, new Action0() {
-                    @Override
-                    public void call() {
-                        //mView.hidLoading();
-                    }
-                });
-        addSubscribe(rxSubscription);
+//        Subscription rxSubscription = RetrofitHelper.getVideoApi().getVideoList(dataId)
+//                .compose(RxUtil.<VideoHttpResponse<VideoRes>>rxSchedulerHelper())
+//                .compose(RxUtil.<VideoRes>handleResult())
+//                .subscribe(new Action1<VideoRes>() {
+//                    @Override
+//                    public void call(final VideoRes res) {
+//                        if (res != null) {
+//                            mView.showContent(res);
+//                            result = res;
+//                            postData();
+////                            insertRecord();
+//                        }
+//                    }
+//                }, new Action1<Throwable>() {
+//                    @Override
+//                    public void call(Throwable throwable) {
+//                        mView.showError("数据加载失败");
+//                    }
+//                }, new Action0() {
+//                    @Override
+//                    public void call() {
+//                        //mView.hidLoading();
+//                    }
+//                });
+//        addSubscribe(rxSubscription);
     }
 
     private void postData() {
