@@ -3,12 +3,6 @@ package com.app.chao.chaoapp;
 import android.app.Application;
 import android.content.Context;
 
-import com.app.chao.chaoapp.utils.RealmHelper;
-
-import io.realm.Realm;
-import io.realm.RealmConfiguration;
-import io.realm.rx.RealmObservableFactory;
-
 /**
  * Created by Chao on 2017/3/13.
  */
@@ -29,13 +23,14 @@ public class App extends Application {
     }
 
     private void initRealm() {
-        RealmConfiguration realmConfiguration = new RealmConfiguration.Builder(this)
-                .name(RealmHelper.DB_NAME)
-                .schemaVersion(1)
-                .rxFactory(new RealmObservableFactory())
-                .deleteRealmIfMigrationNeeded()
-                .build();
-        Realm.setDefaultConfiguration(realmConfiguration);
+//        Realm.init(this);
+//        RealmConfiguration realmConfiguration = new RealmConfiguration.Builder()
+//                .name(RealmHelper.DB_NAME)
+//                .schemaVersion(1)
+//                .rxFactory(new RealmObservableFactory())
+//                .deleteRealmIfMigrationNeeded()
+//                .build();
+//        Realm.setDefaultConfiguration(realmConfiguration);
     }
 
 
