@@ -4,6 +4,8 @@ import com.app.chao.chaoapp.base.BasePresenter;
 import com.app.chao.chaoapp.base.BaseView;
 import com.app.chao.chaoapp.bean.VideoRes;
 
+import java.util.List;
+
 /**
  * Created by Chao on 2017/3/14.
  */
@@ -12,15 +14,17 @@ public interface FragmentOneContract {
 
     interface View extends BaseView<Presenter> {
 
-        void showContent(VideoRes videoRes);
+        void showContent(List<VideoRes> videoRes);
 
         void refreshFaild(String msg);
 
-        void stopBanner(boolean stop);
+        void showBanner(List<VideoRes> videoRes);
     }
 
     interface Presenter extends BasePresenter {
-        void onRefresh();
+        void showContent(int page);
+
+        void showBanner();
     }
 
 }
