@@ -1,7 +1,7 @@
 package com.app.chao.chaoapp.contract.impl;
 
 import com.app.chao.chaoapp.base.RxPresenter;
-import com.app.chao.chaoapp.bean.HomeVideoData;
+import com.app.chao.chaoapp.bean.VideoRes;
 import com.app.chao.chaoapp.contract.ActivityVideoListContract;
 import com.app.chao.chaoapp.net.RetrofitHelper;
 import com.app.chao.chaoapp.net.VideoHttpResponse;
@@ -56,12 +56,12 @@ public class ActivityVideoSearchPresenter extends RxPresenter implements Activit
 
     @Override
     public void getVideoHomeData() {
-        Subscription rxSubscription = RetrofitHelper.getVideoApi().getVideoHomeData()
-                .compose(RxUtil.<VideoHttpResponse<List<HomeVideoData>>>rxSchedulerHelper())
-                .compose(RxUtil.<List<HomeVideoData>>handleResult())
-                .subscribe(new Action1<List<HomeVideoData>>() {
+        Subscription rxSubscription = RetrofitHelper.getVideoApi().getVideoBanner()
+                .compose(RxUtil.<VideoHttpResponse<List<VideoRes>>>rxSchedulerHelper())
+                .compose(RxUtil.<List<VideoRes>>handleResult())
+                .subscribe(new Action1<List<VideoRes>>() {
                     @Override
-                    public void call(final List<HomeVideoData> res) {
+                    public void call(final List<VideoRes> res) {
                         if (res != null) {
 
                             if (page == 1) {
