@@ -1,25 +1,24 @@
 package com.app.chao.chaoapp.ui.fragment;
 
 import android.graphics.Color;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.app.chao.chaoapp.R;
 import com.app.chao.chaoapp.adapter.BannerAdapter;
 import com.app.chao.chaoapp.adapter.FragmentOneAdapter;
-import com.app.chao.chaoapp.bean.HomeVideoData;
+import com.app.chao.chaoapp.base.Preconditions;
 import com.app.chao.chaoapp.bean.VideoInfo;
 import com.app.chao.chaoapp.bean.VideoRes;
-import com.app.chao.chaoapp.contract.HomeActivityContract;
-import com.app.chao.chaoapp.contract.impl.HomeActivityPresenter;
+import com.app.chao.chaoapp.contract.FragmentOneContract;
+import com.app.chao.chaoapp.contract.impl.FragmentOnePresenter;
 import com.app.chao.chaoapp.utils.JumpUtil;
-import com.app.chao.chaoapp.utils.RxBus;
-import com.app.chao.chaoapp.utils.RxBusSubscriber;
 import com.app.chao.chaoapp.utils.RxSubscriptions;
 import com.app.chao.chaoapp.utils.ScreenUtil;
 import com.cjj.MaterialRefreshLayout;
@@ -31,14 +30,13 @@ import com.jude.rollviewpager.hintview.IconHintView;
 import java.util.List;
 
 import butterknife.BindView;
-import butterknife.ButterKnife;
 import rx.Subscription;
 
 /**
  * Created by Chao on 2017/3/13.
  */
 
-public class TabFragmentOne extends BaseFragment implements HomeActivityContract.View {
+public class TabFragmentOne extends BaseFragment implements FragmentOneContract.View {
     @BindView(R.id.refresh)
     MaterialRefreshLayout materialRefreshLayout;
     @BindView(R.id.recyclerView)
@@ -50,7 +48,8 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
     View headerView;
     FragmentOneAdapter adapter;
     Subscription mRxSub;
-    HomeActivityPresenter mPresenter;
+    int page = 1;
+    FragmentOneContract.Presenter mPresenter;
 
     public static TabFragmentOne newInstance() {
         //if (fragment == null)
@@ -66,11 +65,11 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
 
     @Override
     protected void initView(View inflater) {
-        //new FragmentOnePresenter(this);
+        new FragmentOnePresenter(this);
         headerView = LayoutInflater.from(mContext).inflate(R.layout.recommend_header, null);
-        banner = ButterKnife.findById(headerView, R.id.banner);
-        rlGoSearch = ButterKnife.findById(headerView, R.id.rlGoSearch);
-        etSearchKey = ButterKnife.findById(headerView, R.id.etSearchKey);
+        banner = headerView.findViewById(R.id.banner);
+        rlGoSearch = headerView.findViewById(R.id.rlGoSearch);
+        etSearchKey = headerView.findViewById(R.id.etSearchKey);
         banner.setPlayDelay(2000);
         recyclerView.setAdapter(adapter = new FragmentOneAdapter(getContext()));
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
@@ -88,7 +87,6 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
         });
         //recyclerView.setErrorView(R.layout.view_error);
         //webView.loadUrl("http://www.youku.com");
-        getEvent();
         listener();
     }
 
@@ -105,7 +103,7 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
             @Override
             public void onRefresh(final MaterialRefreshLayout materialRefreshLayout) {
                 //RxBus.getDefault().postSticky("Refresh");
-                mPresenter.getVideoHomeData();
+                mPresenter.showContent(1);
             }
 
             @Override
@@ -115,55 +113,10 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
 
             @Override
             public void onRefreshLoadMore(final MaterialRefreshLayout materialRefreshLayout) {
-                //mPresenter.loadMore();
+                mPresenter.showContent(page + 1);
             }
         });
         //materialRefreshLayout.autoRefresh();
-    }
-
-    private void getEvent() {
-        mPresenter = new HomeActivityPresenter(this);
-        mPresenter.getVideoHomeData();
-        //RxSubscriptions.remove(mRxSub);
-        mRxSub = RxBus.getDefault().toObservableSticky(VideoRes.class)
-                .subscribe(new RxBusSubscriber<VideoRes>() {
-                    @Override
-                    protected void onEvent(final VideoRes videoRes) {
-                        if (videoRes != null) {
-                            adapter.clear();
-                            for (int i = 1; i < videoRes.list.size(); i++) {
-                                if (videoRes.list.get(i).title.equals("精彩推荐")) {
-                                    //adapter.addAll(videoRes.list.get(i).childList);
-                                    break;
-                                }
-                            }
-                            for (int i = 1; i < videoRes.list.size(); i++) {
-                                if (videoRes.list.get(i).title.equals("免费推荐")) {
-                                    recommend = videoRes.list.get(i).childList;
-                                    break;
-                                }
-                            }
-                            if (adapter.getHeaderCount() == 0) {
-                                adapter.addHeader(new RecyclerArrayAdapter.ItemView() {//Banner图
-                                    @Override
-                                    public View onCreateView(ViewGroup parent) {
-                                        banner.setHintView(new IconHintView(getContext(), R.mipmap.ic_page_indicator_focused, R.mipmap.ic_page_indicator, ScreenUtil.dip2px(getContext(), 10)));
-                                        banner.setHintPadding(0, 0, 0, ScreenUtil.dip2px(getContext(), 8));
-                                        //banner.setAdapter(new BannerAdapter(getContext(), videoRes.list.get(0).childList));
-                                        return headerView;
-                                    }
-
-                                    @Override
-                                    public void onBindView(View headerView) {
-
-                                    }
-                                });
-                            }
-                            close();
-                        }
-                    }
-                });
-        RxSubscriptions.add(mRxSub);
     }
 
     private void close() {
@@ -175,12 +128,6 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
         super.onDestroy();
         RxSubscriptions.remove(mRxSub);
     }
-
-    //    @Override
-//    public void setPresenter(FragmentOneContract.Presenter presenter) {
-//        mPresenter = Preconditions.checkNotNull(presenter);
-//        presenter.onRefresh();
-//    }
 
     @Override
     public void onResume() {
@@ -195,25 +142,26 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
     }
 
     @Override
-    public void setPresenter(HomeActivityContract.Presenter presenter) {
+    public void showContent(List<VideoRes> videoRes) {
+        adapter.addAll(videoRes);
+        materialRefreshLayout.finishRefresh();
+        materialRefreshLayout.finishRefreshLoadMore();
+    }
+
+    @Override
+    public void refreshFaild(String msg) {
 
     }
 
     @Override
-    public void showContent(VideoRes videoRes) {
-
-    }
-
-    @Override
-    public void showBanner(final List<HomeVideoData> videoRes) {
-        adapter.addAll(videoRes.subList(3, videoRes.size()));
+    public void showBanner(final List<VideoRes> videoRes) {
         if (adapter.getHeaderCount() == 0) {
             adapter.addHeader(new RecyclerArrayAdapter.ItemView() {//Banner图
                 @Override
                 public View onCreateView(ViewGroup parent) {
                     banner.setHintView(new IconHintView(getContext(), R.mipmap.ic_page_indicator_focused, R.mipmap.ic_page_indicator, ScreenUtil.dip2px(getContext(), 10)));
                     banner.setHintPadding(0, 0, 0, ScreenUtil.dip2px(getContext(), 8));
-                    banner.setAdapter(new BannerAdapter(getContext(), videoRes.subList(0, 3)));
+                    banner.setAdapter(new BannerAdapter(getContext(), videoRes));
                     return headerView;
                 }
 
@@ -224,5 +172,12 @@ public class TabFragmentOne extends BaseFragment implements HomeActivityContract
             });
         }
         close();
+    }
+
+    @Override
+    public void setPresenter(FragmentOneContract.Presenter presenter) {
+        mPresenter = Preconditions.checkNotNull(presenter);
+        mPresenter.showBanner();
+        mPresenter.showContent(1);
     }
 }
