@@ -1,18 +1,17 @@
 package com.app.chao.chaoapp.ui.fragment;
 
 import android.os.Bundle;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.app.chao.chaoapp.R;
 import com.app.chao.chaoapp.adapter.RelatedAdapter;
-import com.app.chao.chaoapp.bean.HomeVideoData;
 import com.app.chao.chaoapp.bean.VideoRes;
 import com.app.chao.chaoapp.contract.impl.VideoInfoPresenter;
-import com.app.chao.chaoapp.utils.JumpUtil;
 import com.app.chao.chaoapp.utils.ScreenUtil;
 import com.app.chao.chaoapp.utils.StringUtils;
 import com.app.chao.chaoapp.view.TextViewExpandableAnimation;
@@ -22,10 +21,7 @@ import com.jude.easyrecyclerview.decoration.SpaceDecoration;
 import org.simple.eventbus.EventBus;
 import org.simple.eventbus.Subscriber;
 
-import java.io.Serializable;
-
 import butterknife.BindView;
-import butterknife.ButterKnife;
 
 public class VideoIntroFragment extends BaseFragment {
 
@@ -42,7 +38,7 @@ public class VideoIntroFragment extends BaseFragment {
     }
 
 
-    public static VideoIntroFragment newInstance(HomeVideoData videoInfo) {
+    public static VideoIntroFragment newInstance(VideoRes videoInfo) {
         Bundle args = new Bundle();
         args.putSerializable("video", videoInfo);
         VideoIntroFragment fragment = new VideoIntroFragment();
@@ -55,7 +51,7 @@ public class VideoIntroFragment extends BaseFragment {
         EventBus.getDefault().register(this);
 
         headerView = LayoutInflater.from(mContext).inflate(R.layout.intro_header, null);
-        tvExpand = ButterKnife.findById(headerView, R.id.tv_expand);
+        tvExpand = headerView.findViewById(R.id.tv_expand);
         recyclerView.setAdapter(adapter = new RelatedAdapter(getContext()));
         GridLayoutManager gridLayoutManager = new GridLayoutManager(mContext, 3);
         gridLayoutManager.setSpanSizeLookup(adapter.obtainGridSpanSizeLookUp(3));
@@ -83,7 +79,7 @@ public class VideoIntroFragment extends BaseFragment {
 
             }
         });
-        HomeVideoData video = (HomeVideoData) getArguments().getSerializable("video");
+        VideoRes video = (VideoRes) getArguments().getSerializable("video");
         setData(video);
     }
 
@@ -94,10 +90,10 @@ public class VideoIntroFragment extends BaseFragment {
     }
 
     @Subscriber(tag = VideoInfoPresenter.Refresh_Video_Info)
-    public void setData(HomeVideoData videoInfo) {
-        String dir = "导演：" + StringUtils.removeOtherCode("Mr·Zhang");
-        String act = "主演：" + StringUtils.removeOtherCode("Mr·Zhang");
-        String des = dir + "\n" + act + "\n" + "简介：" + StringUtils.removeOtherCode(videoInfo.getDescription());
+    public void setData(VideoRes videoInfo) {
+        String dir = "导演：" + StringUtils.removeOtherCode(videoInfo.getToStar());
+        String act = "主演：" + StringUtils.removeOtherCode(videoInfo.getPerformer());
+        String des = dir + "\n" + act + "\n" + "简介：" + StringUtils.removeOtherCode(videoInfo.getVideoDescribe());
         tvExpand.setText(des);
 //        if (videoInfo.list.size() > 1)
 //            adapter.addAll(videoInfo.list.get(1).childList);
