@@ -46,10 +46,10 @@ public class RxUtil {
                 return httpResponseObservable.flatMap(new Func1<VideoHttpResponse<T>, Observable<T>>() {
                     @Override
                     public Observable<T> call(VideoHttpResponse<T> videoHttpResponse) {
-                        if (videoHttpResponse.getCode() == 200) {
+                        if (videoHttpResponse.isSuccess()) {
                             return createData(videoHttpResponse.getResult());
-                        } else if (!TextUtils.isEmpty(videoHttpResponse.getMessage())) {
-                            return Observable.error(new ApiException("*" + videoHttpResponse.getMessage()));
+                        } else if (!TextUtils.isEmpty(videoHttpResponse.getMsg())) {
+                            return Observable.error(new ApiException("*" + videoHttpResponse.getMsg()));
                         } else {
                             return Observable.error(new ApiException("*" + "服务器返回error"));
                         }
