@@ -27,7 +27,7 @@ public class StatusBarUtils {
             window.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS
                     /*| WindowManager.LayoutParams.FLAG_TRANSLUCENT_NAVIGATION*/);
             window.getDecorView().setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                   /* | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION*/
+                    /* | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION*/
                     | View.SYSTEM_UI_FLAG_LAYOUT_STABLE);
             window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
             window.setStatusBarColor(Color.TRANSPARENT);
